+++ .github/workflows/L1-tests.yml
@@ -435,8 +435,8 @@ jobs:
           -DPLUGIN_AVINPUT=ON
           -DPLUGIN_HDMIINPUT=ON
           -DPLUGIN_HDCPPROFILE=ON
-          -DPLUGIN_HDMICECSOURCE=ON
-          -DPLUGIN_HDMICECSINK=ON
+          -DPLUGIN_HDMICECSOURCE=OFF
+          -DPLUGIN_HDMICECSINK=OFF
           -DUSE_THUNDER_R4=ON
           -DHIDE_NON_EXTERNAL_SYMBOLS=OFF
           &&
@@ -513,8 +513,8 @@ jobs:
           -DPLUGIN_AVINPUT=ON
           -DPLUGIN_HDMIINPUT=ON
           -DPLUGIN_HDCPPROFILE=ON
-          -DPLUGIN_HDMICECSOURCE=ON
-          -DPLUGIN_HDMICECSINK=ON
+          -DPLUGIN_HDMICECSOURCE=OFF
+          -DPLUGIN_HDMICECSINK=OFF
           -DRDK_SERVICES_L1_TEST=ON
           -DUSE_THUNDER_R4=ON
           -DHIDE_NON_EXTERNAL_SYMBOLS=OFF
