+++ HdmiCecSink/HdmiCecSink.config
@@ -5,3 +5,13 @@ set (callsign "org.rdk.HdmiCecSink")
 if(PLUGIN_HDMICECSINK_STARTUPORDER)
 set (startuporder ${PLUGIN_HDMICECSINK_STARTUPORDER})
 endif()
+
+
+map()
+   key(root)
+   map()
+       kv(mode ${PLUGIN_HDMICECSOURCE_MODE})
+       kv(locator lib${PLUGIN_IMPLEMENTATION}.so)
+   end()
+end()
+ans(configuration)
