+++ HdmiCecSource/HdmiCecSourceImplementation.h
@@ -68,11 +68,11 @@ namespace WPEFramework {
 	        void process (const InActiveSource &msg, const Header &header);
 	        void process (const ImageViewOn &msg, const Header &header);
 	        void process (const TextViewOn &msg, const Header &header);
-	        void process (const RequestActiveSource &msg, const Header &header);
+	        void process (const RequestActiveSourceMessage &msg, const Header &header);
 	        void process (const Standby &msg, const Header &header);
 	        void process (const GetCECVersion &msg, const Header &header);
 	        void process (const CECVersion &msg, const Header &header);
-	        void process (const SetMenuLanguage &msg, const Header &header);
+	        void process (const SetMenuLanguageMessage &msg, const Header &header);
 	        void process (const GiveOSDName &msg, const Header &header);
 	        void process (const GivePhysicalAddress &msg, const Header &header);
 	        void process (const GiveDeviceVendorID &msg, const Header &header);
